Prototype ready for quark client work.
The Client class is rather messy. There is probably a better abstraction for that than a near 1:1 port from Bozzo's javascript directory.

A lot of the stuff in the model package will change in the next 1 to 2 weeks but this is a working stub to get started with the quark impl.