Added hub-config and hub-agent projects.
The Hub Agent is responsible for reacting to Hub Manager
notifications and launching a Hub after the Hub has been
properly configured for the tenant.

The hub-config is some utility classes for loading and
parsing YAML config files. Eventually it will be used
so that all Hub projects may share common configuration.

Also added the initial cut of the Packer.io file which is
used to bake the Hub AMI